~ WIP json codec
